Update the tests with the new API.
Tests are broken, but it's normal, at this stage :-)
