+++ kinto_client/tests/support.py
@@ -1,0 +1,23 @@
+import unittest2 as unittest
+
+import mock
+
+
+def mock_response(session, data=None, permissions=None, headers=None,
+                  error=False):
+    data = data or {}
+    permissions = permissions or {}
+    headers = headers or {}
+    info = {'data': data, 'permissions': permissions}
+    if error:
+        session.request.side_effect = ValueError
+    else:
+        session.request.return_value = (info, headers)
+
+
+def get_record(id=None, data=None, permissions=None):
+    record = mock.MagicMock()
+    record.id = id or '1234'
+    record.data = data or {'foo': 'bar'}
+    record.permissions = permissions or {'read': ['Niko', 'Mat']}
+    return record
